+++ components/postgres/src/examplePayloads.ts
@@ -1,4 +1,4 @@
-import type { TriggerPayload } from "@prismatic-io/spectral";
+import type { pollTablePerform } from "./triggers/pollTablePerform";
 export const queryExamplePayload = {
   data: [
     {
@@ -19,9 +19,7 @@ export const queryExamplePayload = {
     },
   ],
 };
-export const pollTableExamplePayload: {
-  payload: TriggerPayload;
-} = {
+export const pollTableExamplePayload = {
   payload: {
     headers: {},
     queryParameters: {},
@@ -74,5 +72,7 @@ export const pollTableExamplePayload: {
     flow: { id: "testFlowId", name: "Test Flow", stableId: "testFlowStableId" },
     startedAt: "2024-01-15T00:00:00.000Z",
     globalDebug: false,
+    paginationState: undefined,
   },
-};
+  polledNoChanges: false,
+} satisfies Awaited<ReturnType<typeof pollTablePerform>>;