+++ components/postgres/src/outputSchemas/index.ts
@@ -1,0 +1,1 @@
+export { queryOutputSchema } from "./query";